Prefixed actions with "Gitflow:"
...to make action naming consistent with other plugins and to make
searching for gitflow feature easier from the action or global search.

Other plugins typically apply a prefix (like "git:" or "perforce:") if the action is specific to a single plugin. Prefixes are only omitted when they contribute to a global action (like commit, which exists for multiple VCS).
Also added ellipsis (...) to the actions to indicate that further information is required and prompted for in the form of a dialog after the action is selected. See http://ux.stackexchange.com/questions/30420/what-are-the-rules-of-using-ellipsis-on-buttons for instance.